rename MinimalApi project
